New Implementation File
Initially, contestants were going to be given multiple .csv files as input for developing their program. This has since changed to JSON files since I ended up taking more time on trying to process the input files, rather than actually working on the problem.

JSON_Solution.cpp is going to be the solution to the problem which contestants will be facing.