Traffic Simulation and Late Deliveries
Traffic simulation between 07:00 to 10:00 and 18:00 to 20:00 implemented. Edge cases for new system addressed. Turns out, it is impossible to deliver packages to North Bay if the deadline is on monday. This is because no matter what, the driver will have to drive longer than 14 hours to arrive by the deadline. For this reason, late deliveries were logged.

Narrowing conversions handled. Minor bug fixes and compiler warnings addressed.